Add support for a single carousel item
I manually merged this from:

https://github.com/willowtreeapps/ouroboros/pull/41

Currently, when we try to run the carousel with a single item we’ll crash with the following precondition:

```
precondition(count >= buffer, "Ouroboros requires at least twice the number of items per page to work properly. For best results: a number that is evenly divisible by the number of items per page.")
```

Infinite scrolling is essentially the reason it didn’t work before.  There’s no way to do that with a single item. He added support for non-infinite scroll.

It took a lot of work to manually merge in his changes with Fredricks and with what was different on the origin repo.  There were too many conflicts to do it automatically.