Fix index exception showing the regular infinite scroll
We were getting an exception relating to indexes when doing the regular infinite scroll.  The main cause was actually just a block that got repeated during a big merge conflict:

```
        if focusHeading == .left && to.item < buffer {
            jumping = true
            currentlyFocusedItem += count
        }

        if focusHeading == .right && to.item >= buffer + count {
            jumping = true
            currentlyFocusedItem -= count
        }
```

Should have gone _inside_ the `if shouldEnableInfiniteScroll` section only.  It was actually duplicated in there.  This is custom code that the support for few items added, and it was accidentally in twice.

I also redid more of the class trying to pull in the exact formatting from the PR we were using.